Frontend -  Make a resusable function and camelCase sythaxe file rename
