+++ frontend/src/Components/Dashboards/ChartsAdmin.jsx
@@ -8,6 +8,9 @@ DashboardVolunteers Component | Capstone App (Pursuit Volunteer Mgr)
 import React, { useState, useEffect } from 'react';
 import Chart from '../Chart';
 
+import yearRange from '../../yearRangeFormatting';
+
+
 export default function Charts(props) {
   const { chartData } = props;
   const [chartInterval, setChartInterval] = useState([]);
@@ -17,32 +20,10 @@ export default function Charts(props) {
 
   useEffect(() => {
     if (chartData !== undefined) {
-      let month = new Date().getMonth() + 2;
-      let year = new Date().getFullYear() - 1;
-      const datesArr = [];
+      const datesArr = yearRange();
       const hoursArr = [];
       const eventsArr = [];
       const volunteersArr = [];
-
-      for (let i = 0; i < 12; i++) {
-        let date = '';
-        if (month <= 12) {
-          if (month < 10) {
-            date = '0' + month + '-' + year;
-          } else {
-            date = month + '-' + year;
-          }
-          month += 1;
-        } else {
-          if (month < 22) {
-            date = '0' + (month - 12) + '-' + (year + 1);
-          } else {
-            date = (month - 12) + '-' + (year + 1);
-          }
-          month += 1;
-        }
-        datesArr.push(date);
-      }
 
       for (let date of datesArr) {
         let h = null;
